Create a Task interface and refactor JSONification
After this commit, what were previously called Tasks are now called
DefiniteTasks, because they have a definite due date. They implement the
new Task interface, which has methods that allowed it to be a largely
drop-in replacement for DefiniteTask. This means that, in the future,
other types of Tasks (see #17) can be added easily.

Additionally, I rewrote the in-file layout to use a different structure
that sorts different types of tasks apart from one another. This will
allow for cleaner storage once there are multiple types of tasks.