Add eventual todos, which are a work-in-progress
They are added using the "eventual" command, the syntax of which is
simillar to definite todos, except it takes a duration, such as 8h,
rather than an exact date. This is used to calculate an urgency ("Nice")
value.

Eventual todos are very much a work-in-progress, and really need a
better relative time library to work better.

Ref #17