Make use of my relative time formatting library
TaskToGo's relative time formatting will, with this commit, begin to
improve along with the mentioned library. It makes use of the
highest-available function for formatting the due date exactly in a more
human-readable format, such as "tomorrow" or "today". When the library
gains support for strings like "Monday" or "Next Monday", then those
will be used automatically, as well.

Ref #14